// src/constants/generatedSkinCatalog.ts
export interface GeneratedSkinCatalogItem {
  id: number;
  name: string;
  brawlerId: number;
  brawlerName: string;
  brawlerNameKo: string;
  rarity: string;
  gems: number;
  bling: number;
  coins: number;
  isDefault: boolean;
  isCatalogReleased: boolean;
}